[security] Google Recaptcha - Integrate recaptcha for delete/upload key

+++ SSHConfigurator/Controllers/HomeController.cs
@@ -14,6 +14,7 @@ using SSHConfigurator.Domain;
 using SSHConfigurator.Models;
 using SSHConfigurator.Services;
 using SSHConfigurator.ViewModels;
+using Vereyon.Web;
 using ErrorViewModel = SSHConfigurator.ViewModels.ErrorViewModel;
 
 namespace SSHConfigurator.Controllers
@@ -25,15 +26,17 @@ namespace SSHConfigurator.Controllers
         private readonly SignInManager<THUMember> signInManager;
         private readonly IKeyStorageService keyStorageService;
         private readonly IWebHostEnvironment iHostingEnvironment;
+        private readonly GoogleRecaptchaService recaptchaService;
 
         public HomeController(ILogger<HomeController> logger, UserManager<THUMember> userManager, SignInManager<THUMember> signInManager,
-                              IKeyStorageService keyStorageService, IWebHostEnvironment IHostingEnvironment)
+                              IKeyStorageService keyStorageService, IWebHostEnvironment IHostingEnvironment, GoogleRecaptchaService recaptchaService)
         {
             _logger = logger;
             this.userManager = userManager;
             this.signInManager = signInManager;
             this.keyStorageService = keyStorageService;
             iHostingEnvironment = IHostingEnvironment;
+            this.recaptchaService = recaptchaService;
         }
 
         [HttpGet]
@@ -49,8 +52,18 @@ namespace SSHConfigurator.Controllers
         }
             
         [HttpPost]
-        public async Task<IActionResult> DeleteKey()
+        public async Task<IActionResult> DeleteKey(HomeViewModel deleteViewModel)
         {
+            // Google Recaptcha Verification
+            var googleRecaptcha = await recaptchaService.ReceiveVerificationAsync(deleteViewModel.Token);
+
+            // If verification failed, sign out user.
+            if (!googleRecaptcha.Success)
+            {
+                await signInManager.SignOutAsync();
+                return RedirectToAction("login", "account");
+            }
+
             await keyStorageService.DeletePublicKeyAsync(User.Identity.Name);
             return RedirectToAction("Index");
         }
@@ -64,6 +77,15 @@ namespace SSHConfigurator.Controllers
         [HttpPost]
         public async Task<IActionResult> UploadKey(UploadKeyViewModel uploadKeyViewModel)
         {
+            // Google Recaptcha Verification
+            var googleRecaptcha = await recaptchaService.ReceiveVerificationAsync(uploadKeyViewModel.Token);
+
+            // If verification failed, sign out user.
+            if (!googleRecaptcha.Success)
+            {
+                await signInManager.SignOutAsync();
+                return RedirectToAction("login", "account");
+            }
             if (ModelState.IsValid)
             {
                 // temporarily store the key in the www/temp-keys folder
